Synchronize with GitHub Repackage all

+++ Template.DB/PowerShell/NuGetDbPacker.psm1
@@ -153,14 +153,17 @@ function Publish-DbPackage {
     $configPath = [IO.Path]::ChangeExtension($ProjectPath, '.nuget.config')
     $projFolder = Split-Path $ProjectPath -Resolve
     $nugetFolder = [IO.Path]::Combine($projFolder, 'NuGet')
-    $settings = Import-NuGetSettings -NugetConfigPath $configPath -SolutionPath $SolutionPath
-    $id = $settings.nugetSettings.Id
-    $version = $settings.nugetSettings.version
-    if (-not (Test-NuGetVersionExists -Id $id -Version $version)) {
-        $nugetPackage = [IO.Path]::Combine($nugetFolder, "$id.$version.nupkg")
-        Initialize-DbPackage -ProjectPath $ProjectPath -SolutionPath $SolutionPath
-		Publish-NuGetPackage -PackagePath $nugetPackage
-		Remove-NugetFolder $nugetFolder
+    if (Test-Path $configPath)
+    {
+        $settings = Import-NuGetSettings -NugetConfigPath $configPath -SolutionPath $SolutionPath
+        $id = $settings.nugetSettings.Id
+        $version = $settings.nugetSettings.version
+        if (-not (Test-NuGetVersionExists -Id $id -Version $version)) {
+            $nugetPackage = [IO.Path]::Combine($nugetFolder, "$id.$version.nupkg")
+            Initialize-DbPackage -ProjectPath $ProjectPath -SolutionPath $SolutionPath
+            Publish-NuGetPackage -PackagePath $nugetPackage
+            Remove-NugetFolder $nugetFolder
+        }
     }
 }
 
@@ -252,6 +255,7 @@ function Initialize-DbPackage
 	$nugetSettings = Import-NuGetSettings -NugetConfigPath $configPath -SolutionPath $SolutionPath
 
 	Initialize-Package -ProjectPath $ProjectPath -NugetSettings $nugetSettings
+	mkdir "$nugetPath\content\Databases" | Out-Null
 	Import-NuGetDb -ProjectPath $ProjectPath -ProjDbFolder "$projectFolder\Databases" -NugetDbFolder "$nugetPath\content\Databases" -NugetSpecPath "$nugetPath\Package.nuspec"
 	Compress-Package -NugetPath $nugetPath
 }
